refactor: changed users.js -> user.js. added special handling for CustomException in routes.
