+++ app.js
@@ -1,10 +1,10 @@
 import express from 'express';
 const app = express();
-// import configRoutes from './routes/index.js';
+import configRoutes from './routes/index.js';
 
 app.use(express.json());
 
-// configRoutes(app);
+configRoutes(app);
 
 app.listen(3000, async () => {
   console.log("We've now got a server!");
