+++ routes/index.js
@@ -1,12 +1,11 @@
-import userRoutes from './users.js';
-import * as userData from '../data/users.js';
+import userRoutes from './user.js';
 
 const constructorMethod = (app) => {
-    app.use('/', userRoutes);
-   
-    app.use('*', (req, res) => {
-      res.status(404).json({ error: 'Not found' });
-    });
-  };
+  app.use('/user', userRoutes);
+ 
+  app.use('*', (req, res) => {
+    res.status(404).json({ error: 'Not found' });
+  });
+};
   
-  export default constructorMethod;
+export default constructorMethod;
